Remove unnecessary boxing unboxing

+++ stream/common/src/main/java/org/apache/bookkeeper/common/coder/VarIntCoder.java
@@ -50,7 +50,7 @@ public class VarIntCoder implements Coder<Integer> {
         ByteBufOutputStream output = new ByteBufOutputStream(buf);
 
         try {
-            VarInt.encode(value.intValue(), output);
+            VarInt.encode(value, output);
         } catch (IOException e) {
             throw new IllegalStateException("Failed to encode integer '" + value + "' into the provided buffer", e);
         }
